# backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Stellar Privacy Analytics API
  description: |
    Privacy-first analytics API for the Stellar ecosystem that provides secure data analysis with built-in differential privacy and comprehensive audit logging.
    
    ## Features
    - 🔐 JWT-based authentication with role-based access control
    📊 Privacy-preserving analytics with differential privacy
    🔍 Advanced query capabilities with noise injection
    📋 Comprehensive audit logging
    🚀 Real-time data processing
    🛡️ Enterprise-grade security
    
    ## Authentication
    This API uses JWT tokens for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    API requests are rate-limited based on your subscription tier:
    - Basic: 100 requests per 15 minutes
    - Premium: 500 requests per 15 minutes  
    - Enterprise: 2000 requests per 15 minutes
    
    ## Privacy Budget
    Each query consumes privacy budget measured in epsilon (ε). The system automatically adds noise to protect individual privacy while maintaining statistical accuracy.
  version: 1.0.0
  contact:
    name: Stellar Privacy Analytics Team
    email: support@stellar-privacy.com
    url: https://stellar-privacy.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.stellar-privacy.com/v1
    description: Production server
  - url: https://staging-api.stellar-privacy.com/v1
...
paths:
...
  /query/schema:
    get:
      tags:
        - Query
      summary: Get data schema
      description: Retrieve available data fields and their types
      operationId: getDataSchema
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Schema retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSchemaResponse'
  /query/validate:
    post:
      tags:
        - Query
      summary: Validate query
      description: Validate a privacy query structure without executing it
      operationId: validateQuery
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateQueryRequest'
      responses:
        '200':
          description: Query validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateQueryResponse'
  /query/privacy-cost:
    post:
      tags:
        - Query
      summary: Calculate privacy cost
      description: Calculate the privacy budget cost for a query
      operationId: calculatePrivacyCost
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrivacyCostRequest'
      responses:
        '200':
          description: Privacy cost calculated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivacyCostResponse'
  /query/execute:
    post:
      tags:
        - Query
      summary: Execute privacy query
      description: Execute a privacy-preserving query with noise injection
      operationId: executeQuery
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteQueryRequest'
      responses:
        '200':
          description: Query executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecuteQueryResponse'
  /query/favorites:
    get:
      tags:
        - Query
      summary: Get favorite queries
      description: Retrieve user's saved favorite queries
      operationId: getFavoriteQueries
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: query
          description: User ID (optional, defaults to current user)
          schema:
            type: string
      responses:
        '200':
          description: Favorite queries retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteQueriesResponse'

    post:
      tags:
        - Query
      summary: Save favorite query
      description: Save a query as a favorite for future use
      operationId: saveFavoriteQuery
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveFavoriteRequest'
      responses:
        '201':
          description: Query saved as favorite
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveFavoriteResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
  schemas:
...
    DataSchemaResponse:
      type: object
      properties:
        schema:
          type: array
          items:
            $ref: '#/components/schemas/DataField'
        message:
          type: string
          example: Data schema retrieved successfully
    DataField:
      type: object
      properties:
        id:
          type: string
          example: user_id
        name:
          type: string
          example: User ID
        type:
          type: string
          enum: [string, number, date, boolean]
          example: string
        description:
          type: string
          example: Unique identifier for users
    PrivacyQuery:
      type: object
      required:
        - steps
      properties:
        steps:
          type: array
          items:
            $ref: '#/components/schemas/QueryStep'
          description: Array of query processing steps
    QueryStep:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum: [select, filter, aggregate, group]
          example: filter
        field:
          type: string
          example: transaction_amount
        operator:
          type: string
          enum: [eq, gt, lt, gte, lte, contains, in]
          example: gt
        value:
          description: Filter value (varies by type)
          example: 100
        aggregation:
          type: string
          enum: [count, sum, average, min, max]
          example: sum
    ValidateQueryRequest:
      type: object
      required:
        - query
      properties:
        query:
          $ref: '#/components/schemas/PrivacyQuery'
    ValidateQueryResponse:
      type: object
      properties:
        valid:
          type: boolean
          example: true
        errors:
          type: array
          items:
            type: string
          example: []
        message:
          type: string
          example: Query is valid
    PrivacyCostRequest:
      type: object
      required:
        - query
      properties:
        query:
          $ref: '#/components/schemas/PrivacyQuery'
    PrivacyCostResponse:
      type: object
      properties:
        cost:
          type: number
          example: 2.5
        breakdown:
          type: object
          properties:
            base:
              type: number
              example: 3.0
            final:
              type: number
              example: 2.5
            hasAggregation:
              type: boolean
              example: true
            stepsCount:
              type: integer
              example: 3
        riskLevel:
          type: string
          enum: [low, medium, high]
          example: medium
        epsilon:
          type: string
          example: "0.25"
        message:
          type: string
          example: Privacy cost calculated successfully
    ExecuteQueryRequest:
      type: object
      required:
        - query
      properties:
        query:
          $ref: '#/components/schemas/PrivacyQuery'
        walletSignature:
          type: string
          description: Optional wallet signature for verification
          example: sig_abc123...
    ExecuteQueryResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        jobId:
          type: string
          example: job-def456
        status:
          type: string
          example: completed
        results:
          type: object
          properties:
            summary:
              type: object
            data:
              type: array
              items:
                type: object
        privacyMetrics:
          $ref: '#/components/schemas/PrivacyMetrics'
        executionTime:
          type: integer
          description: Execution time in milliseconds
          example: 2340
        message:
          type: string
          example: Query executed successfully with privacy preservation
    PrivacyMetrics:
      type: object
      properties:
        cost:
          type: number
          example: 2.5
        epsilon:
          type: string
          example: "0.25"
        riskLevel:
          type: string
          enum: [low, medium, high]
          example: medium
        noiseAdded:
          type: boolean
          example: true
        anonymizationStrength:
          type: number
          example: 0.85
    FavoriteQueriesResponse:
      type: object
      properties:
        favorites:
          type: array
          items:
            $ref: '#/components/schemas/FavoriteQuery'
        message:
          type: string
          example: Favorite queries retrieved successfully
    FavoriteQuery:
      type: object
      properties:
        id:
          type: string
          example: fav-123
        name:
          type: string
          example: High Value Transactions
        description:
          type: string
          example: Find transactions over $500
        userId:
          type: string
          example: user-abc123
        createdAt:
          type: string
          format: date-time
        privacyCost:
          type: number
          example: 2.5
        query:
          $ref: '#/components/schemas/PrivacyQuery'
    SaveFavoriteRequest:
      type: object
      required:
        - name
        - query
      properties:
        name:
          type: string
          description: Favorite query name
          example: Monthly Revenue
        description:
          type: string
          description: Query description
          example: Monthly revenue analysis
        query:
          $ref: '#/components/schemas/PrivacyQuery'
        userId:
          type: string
          description: User ID (optional, defaults to authenticated user)
          example: user-abc123
    SaveFavoriteResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        favorite:
          $ref: '#/components/schemas/FavoriteQuery'
        message:
          type: string
          example: Favorite query saved successfully
...