# backend/src/routes/query.ts
import { Router } from "express";
import { body, query } from "express-validator";
import { asyncHandler } from "../middleware/errorHandler";
import {
  AggregationType,
  AnalysisFilter,
} from "@stellar/shared";
import { validateRequest } from "../middleware/validation";

const router = Router();

const queryPayloadRules = [
  body("query").isObject().withMessage("query must be an object"),
  body("query.steps")
    .isArray({ min: 1, max: 500 })
    .withMessage("query.steps must be a non-empty array"),
];

// Mock data schema
const dataSchema = [
  {
    id: "user_id",
    name: "User ID",
    type: "string",
    description: "Unique identifier for users",
  },
  {
    id: "transaction_amount",
    name: "Transaction Amount",
    type: "number",
    description: "Amount in USD",
  },
  {
    id: "transaction_date",
    name: "Transaction Date",
    type: "date",
    description: "Date of transaction",
  },
  {
    id: "user_type",
    name: "User Type",
    type: "string",
    description: "User category",
  },
  {
    id: "is_active",
    name: "Active Status",
    type: "boolean",
    description: "Whether user is active",
  },
];

// Get available data fields and schema
router.get(
  "/schema",
  asyncHandler(async (req, res) => {
    res.json({
      schema: dataSchema,
      message: "Data schema retrieved successfully",
    });
  }),
);

// Validate query structure
router.post(
  "/validate",
  [...queryPayloadRules, validateRequest],
  asyncHandler(async (req, res) => {
    const { query } = req.body;

    if (!query || !query.steps || !Array.isArray(query.steps)) {
      return res.status(400).json({
        valid: false,
        errors: ["Query must have steps array"],
        message: "Invalid query structure",
      });
    }

    const errors: string[] = [];
    const steps = query.steps;

    // Check if at least one select step exists
    const hasSelect = steps.some((step) => step.type === "select");
    if (!hasSelect) {
      errors.push("Query must include at least one field selection");
    }

    // Validate each step
    steps.forEach((step: any, index: number) => {
      if (!step.type) {
        errors.push(`Step ${index + 1}: Missing type`);
        return;
      }

      switch (step.type) {
        case "select":
          if (!step.field) {
            errors.push(`Step ${index + 1}: Select step missing field`);
          }
          break;
        case "filter":
          if (!step.field || !step.operator || step.value === undefined) {
            errors.push(
              `Step ${index + 1}: Filter step incomplete (missing field, operator, or value)`,
            );
          }
          break;
        case "aggregate":
          if (!step.field || !step.aggregation) {
            errors.push(
              `Step ${index + 1}: Aggregate step incomplete (missing field or aggregation)`,
            );
          }
          break;
        case "group":
          if (!step.field) {
            errors.push(`Step ${index + 1}: Group step missing field`);
          }
          break;
        default:
          errors.push(`Step ${index + 1}: Unknown step type '${step.type}'`);
      }
    });

    const isValid = errors.length === 0;

    res.json({
      valid: isValid,
      errors,
      message: isValid ? "Query is valid" : "Query has validation errors",
    });
  }),
);

// Calculate privacy cost for query
router.post(
  "/privacy-cost",
  [...queryPayloadRules, validateRequest],
  asyncHandler(async (req, res) => {
    const { query } = req.body;

    if (!query || !query.steps || !Array.isArray(query.steps)) {
      return res.status(400).json({
        cost: 0,
        errors: ["Invalid query structure"],
        message: "Cannot calculate privacy cost",
      });
    }

    let cost = 0;
    const steps = query.steps;

    steps.forEach((step: any) => {
      switch (step.type) {
        case "select":
          cost += 0.1;
          break;
        case "filter":
          cost += 0.2;
          break;
        case "aggregate":
          cost += 0.5;
          break;
        case "group":
          cost += 0.3;
          break;
      }
    });

    // Apply some privacy-preserving logic
    const hasAggregation = steps.some((s) => s.type === "aggregate");
    if (hasAggregation) {
      cost *= 0.8; // Reduce cost for aggregated queries
    }

    const finalCost = Math.min(cost, 10); // Cap at 10 privacy units

    res.json({
      cost: finalCost,
      breakdown: {
        base: cost,
        final: finalCost,
        hasAggregation,
        stepsCount: steps.length,
      },
      riskLevel: finalCost < 2 ? "low" : finalCost < 5 ? "medium" : "high",
      epsilon: (finalCost * 0.1).toFixed(2),
      message: "Privacy cost calculated successfully",
    });
  }),
);

// Execute privacy-preserved query
router.post(
  "/execute",
  [
    ...queryPayloadRules,
    body("walletSignature")
      .optional({ values: "null" })
      .isString()
      .isLength({ max: 8192 }),
    validateRequest,
  ],
  asyncHandler(async (req, res) => {
    const { query, _walletSignature } = req.body;

    // Validate query first
    const validationResponse: any = await (
      await fetch(
        `${req.protocol}://${req.get("host")}/api/v1/query/validate`,
        {
          method: "POST",
          headers: { "Content-Type": "application/json" },
          body: JSON.stringify({ query }),
        },
      )
    ).json();

    if (!validationResponse.valid) {
      return res.status(400).json({
        success: false,
        errors: validationResponse.errors,
        message: "Query validation failed",
      });
    }

    // Check privacy cost
    const costResponse: any = await (
      await fetch(
        `${req.protocol}://${req.get("host")}/api/v1/query/privacy-cost`,
        {
          method: "POST",
          headers: { "Content-Type": "application/json" },
          body: JSON.stringify({ query }),
        },
      )
    ).json();

    // Simulate query execution with differential privacy
    const mockResults = generateMockResults(query);

    res.json({
      success: true,
      jobId: `job-${Date.now()}`,
      status: "completed",
      results: mockResults,
      privacyMetrics: {
        cost: costResponse.cost,
        epsilon: costResponse.epsilon,
        riskLevel: costResponse.riskLevel,
        noiseAdded: true,
        anonymizationStrength: 0.85,
      },
      executionTime: Math.floor(Math.random() * 5000) + 1000, // 1-6 seconds
      message: "Query executed successfully with privacy preservation",
    });
  }),
);

// Generate mock results for demonstration
function generateMockResults(query: any) {
  const steps = query.steps || [];
  const selectSteps = steps.filter((s: any) => s.type === "select");
  const aggregateSteps = steps.filter((s: any) => s.type === "aggregate");
  const groupSteps = steps.filter((s: any) => s.type === "group");

  // Generate base data
  const baseData = [];
  for (let i = 0; i < 100; i++) {
    const record: any = {
      user_id: `user_${i}`,
      transaction_amount: Math.floor(Math.random() * 1000) + 50,
      transaction_date: new Date(
        Date.now() - Math.random() * 30 * 24 * 60 * 60 * 1000,
      )
        .toISOString()
        .split("T")[0],
      user_type: ["premium", "basic", "enterprise"][
        Math.floor(Math.random() * 3)
      ],
      is_active: Math.random() > 0.3,
    };
    baseData.push(record);
  }

  // Apply filters
  let filteredData = baseData;
  const filterSteps = steps.filter((s: any) => s.type === "filter");

  filterSteps.forEach((filter: any) => {
    switch (filter.operator) {
      case "gt":
        filteredData = filteredData.filter(
          (record: any) => record[filter.field] > filter.value,
        );
        break;
      case "lt":
        filteredData = filteredData.filter(
          (record: any) => record[filter.field] < filter.value,
        );
        break;
      case "eq":
        filteredData = filteredData.filter(
          (record: any) => record[filter.field] === filter.value,
        );
        break;
      case "contains":
        filteredData = filteredData.filter((record: any) =>
          String(record[filter.field]).includes(filter.value),
        );
        break;
    }
  });

  // Apply aggregations
  if (aggregateSteps.length > 0) {
    const results: any[] = [];

    if (groupSteps.length > 0) {
      // Group by logic
      const groups: { [key: string]: any[] } = {};
      const groupField = groupSteps[0].field;

      filteredData.forEach((record) => {
        const key = record[groupField];
        if (!groups[key]) groups[key] = [];
        groups[key].push(record);
      });

      Object.keys(groups).forEach((groupKey) => {
        const groupData = groups[groupKey];
        const result: any = { group: groupKey };

        aggregateSteps.forEach((agg: any) => {
          const values = groupData
            .map((r: any) => r[agg.field])
            .filter((v) => typeof v === "number");

          switch (agg.aggregation) {
            case AggregationType.COUNT:
              result[`${agg.field}_count`] = groupData.length;
              break;
            case AggregationType.SUM:
              result[`${agg.field}_sum`] = values.reduce((a, b) => a + b, 0);
              break;
            case AggregationType.AVERAGE:
              result[`${agg.field}_avg`] =
                values.reduce((a, b) => a + b, 0) / values.length;
              break;
            case AggregationType.MIN:
              result[`${agg.field}_min`] = Math.min(...values);
              break;
            case AggregationType.MAX:
              result[`${agg.field}_max`] = Math.max(...values);
              break;
          }
        });

        results.push(result);
      });
    } else {
      // Simple aggregation without grouping
      const result: any = {};

      aggregateSteps.forEach((agg: any) => {
        const values = filteredData
          .map((r: any) => r[agg.field])
          .filter((v) => typeof v === "number");

        switch (agg.aggregation) {
          case AggregationType.COUNT:
            result[`${agg.field}_count`] = filteredData.length;
            break;
          case AggregationType.SUM:
            result[`${agg.field}_sum`] = values.reduce((a, b) => a + b, 0);
            break;
          case AggregationType.AVERAGE:
            result[`${agg.field}_avg`] =
              values.reduce((a, b) => a + b, 0) / values.length;
            break;
          case AggregationType.MIN:
            result[`${agg.field}_min`] = Math.min(...values);
            break;
          case AggregationType.MAX:
            result[`${agg.field}_max`] = Math.max(...values);
            break;
        }
      });

      results.push(result);
    }

    return {
      summary: {
        totalRecords: baseData.length,
        processedRecords: filteredData.length,
        groups: groupSteps.length,
        aggregations: aggregateSteps.length,
      },
      data: results,
    };
  }

  // Simple select without aggregation
  const selectedFields = selectSteps.map((s: any) => s.field);
  const results = filteredData.map((record) => {
    const selected: any = {};
    selectedFields.forEach((field: string) => {
      selected[field] = record[field];
    });
    return selected;
  });

  return {
    summary: {
      totalRecords: baseData.length,
      processedRecords: filteredData.length,
      selectedFields: selectedFields.length,
    },
    data: results.slice(0, 50), // Limit to 50 for demo
  };
}

// Save favorite query
router.post(
  "/favorites",
  [
    body("name").trim().notEmpty().isLength({ max: 200 }),
    body("description")
      .optional({ values: "null" })
      .trim()
      .isLength({ max: 2000 }),
    body("query").isObject(),
    body("userId").optional({ values: "null" }).trim().isLength({ max: 256 }),
    validateRequest,
  ],
  asyncHandler(async (req, res) => {
    const { name, description, query, userId } = req.body;

    // In a real implementation, this would save to a database
    const favorite = {
      id: `fav-${Date.now()}`,
      name,
      description:
        description || `Query with ${query.steps?.length || 0} steps`,
      query,
      userId: userId || "anonymous",
      createdAt: new Date().toISOString(),
      privacyCost: 0, // Would be calculated
    };

    res.status(201).json({
      success: true,
      favorite,
      message: "Favorite query saved successfully",
    });
  }),
);

// Get favorite queries
router.get(
  "/favorites",
  [query("userId").optional().trim().isLength({ max: 256 }), validateRequest],
  asyncHandler(async (req, res) => {
    const { userId } = req.query;

    // In a real implementation, this would fetch from database
    const favorites = [
      {
        id: "fav-1",
        name: "High Value Transactions",
        description: "Find transactions over $500",
        userId: userId || "anonymous",
        createdAt: new Date().toISOString(),
        privacyCost: 2.5,
        query: {
          steps: [
            { type: "select", field: "transaction_amount" },
            {
              type: "filter",
              field: "transaction_amount",
              operator: "gt",
              value: 500,
            },
            {
              type: "aggregate",
              field: "transaction_amount",
              aggregation: "sum",
            },
          ],
        },
      },
    ];

    res.json({
      favorites,
      message: "Favorite queries retrieved successfully",
    });
  }),
);

export { router as queryRoutes };